Added unparsable url report, moved things to right packages, took iterables from guava
