+++ rules-xml/tests/src/test/java/org/jboss/windup/rules/apps/xml/rules/ValidateXmlFilesRuleProviderWithoutInternetTest.java
@@ -1,5 +1,6 @@
 package org.jboss.windup.rules.apps.xml.rules;
 
+import com.google.common.collect.Iterables;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.forge.arquillian.AddonDependencies;
@@ -13,9 +14,9 @@ import org.jboss.windup.graph.service.GraphService;
 import org.jboss.windup.reporting.model.ClassificationModel;
 import org.jboss.windup.reporting.model.InlineHintModel;
 import org.jboss.windup.reporting.service.InlineHintService;
+import org.jboss.windup.rules.apps.xml.model.XmlFileModel;
 import org.jboss.windup.rules.apps.xml.xml.ValidateXmlFilesRuleProvider;
 import org.jboss.windup.testutil.basics.WindupTestUtilMethods;
-import org.jboss.windup.util.Iterables;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -151,6 +152,28 @@ public class ValidateXmlFilesRuleProviderWithoutInternetTest  extends AbstractXs
                         ClassificationModel.class);
             Iterable<ClassificationModel> classifications = classificationService.findAll();
             Assert.assertEquals(0, Iterables.size(classifications));
+
+        }
+    }
+
+    @Test
+    public void testUnparsableUrl() throws Exception
+    {
+        try (GraphContext context = factory.create())
+        {
+            addFileModel(context, URL_NOT_PARSABLE);
+            initOnlineWindupConfiguration(context);
+
+            List<? extends RuleProvider> ruleProviders = Collections.singletonList(ruleProvider);
+            WindupTestUtilMethods.runOnlyRuleProviders(ruleProviders, context);
+
+            GraphService<InlineHintModel> hintService = new GraphService<>(context,
+                        InlineHintModel.class);
+            Iterable<InlineHintModel> hints = hintService.findAll();
+            Assert.assertEquals(1, Iterables.size(hints));
+
+            final InlineHintModel hint = hints.iterator().next();
+            Assert.assertEquals(XmlFileModel.XSD_URL_NOT_VALID,hint.getTitle());
 
         }
     }
